+++ widget/template/entry.php
@@ -1,0 +1,5 @@
+<li class="item">
+  <input type="checkbox" <?php e($checked, 'checked="checked"') ?> name="panelbar[]" value="<?= $element ?>" />
+  <?= ucfirst($element) ?>
+  <span class="handles"></span>
+</li>